[CD-209] Feat&Refactor: 스크랩 위치 변경 및 정확도 기다리는 부분 모듈화
